requirements/base.txt
jinja2>=2.10
markdown2>=2.3.0
Pygments>=2.0.1
python-swiftclient>=2.2.0
-e .
